refactor: compose NCEP obs sources from mixins in utils_ncep
Replace the _NCEPObsStore protocol and the _NCEPObsSourceBase /
_NNJAObsSourceBase / _NNJAObsStore hierarchy with mixins: NCEPObsRequestMixin
(shared request lifecycle) and NNJAObsArchiveMixin (NNJA S3 transport, cache,
six-hour validation). NNJAObsConv composes both. Fold the request mixin into
utils_ncep and delete the ncep_obs module so shared NCEP code lives in one
util file rather than a module that is not itself a data source.
